refactored flask app
